created CustomChartPanel to separate the chartPanel related methods from PluglnDialogVolumeRenderDualJanelia
